장바구니 구현중

+++ src/main/java/org/teamproject/repositories/BooksRepository.java
@@ -3,5 +3,5 @@ package org.teamproject.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.teamproject.entities.Books;
 
-public interface BooksRepository extends JpaRepository<Books, Long> {
+public interface BooksRepository extends JpaRepository<Books, Integer> {
 }
